% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/beard.R
\name{beard}
\alias{beard}
\title{Score the beard model}
\usage{
beard(input)
}
\arguments{
\item{input}{data passed on as \code{h} to \code{\link{forecast}}}
}
\description{
Simple model with one predictor \code{year}
}
\examples{
mydata <- data.frame(year=c(1921))
   beard(mydata)
}

