% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/beard_model.R
\docType{data}
\name{beard_model}
\alias{beard_model}
\title{Percent of Men With Beards Prediction Model}
\format{a \code{\link{ets}} object}
\source{
https://datamarket.com/data/set/22pd/percent-of-men-with-full-beards-1866-1911#!ds=22pd&display=line
}
\usage{
beard_model
}
\description{
Simple example model that predicts Percent of Men With Beards via Time Series Analysis
}

